pyproject.toml
[project]
name = "metaoct"
version = "1.0.0"
description = "MetaOCT Environment for OpenEnv"
authors = [{ name = "Agent", email = "agent@example.com" }]
dependencies = [
    "openenv-core>=0.0.1",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0"
]

[project.scripts]
server = "server.app:main"
